+++ docs/Dioxus.toml
@@ -6,7 +6,7 @@
 # HTML title tag content
 title = "Dioxus Motion 🚀"
 
-base_path = "dioxus-motion"
+# base_path = "dioxus-motion"
 
 # include `assets` in web platform
 [web.resource]
@@ -22,4 +22,3 @@ script = []
 # Javascript code file
 # serve: [dev-server] only
 script = []
-
